requirements.txt
py-cord==2.4.1
flask
python-dotenv
google-generativeai
yt-dlp
requests
beautifulsoup4
